Move core into library
