+++ script/library/core/GameObject.js
