+++ script/library/core/Room.js
