+++ script/library/core/Sprite.js
