+++ script/library/core/SpriteFrame.js
