+++ script/library/core/SpriteManager.js
