+++ script/library/core/model/actions/spriteFrames.js
