+++ script/library/core/model/actions/spriteSheets.js
