+++ script/library/core/model/actions/sprites.js
