+++ script/library/core/model/gameStateStore.js
@@ -1,4 +1,4 @@
-import createStore from './../../library/store/createStore';
+import createStore from './../../../library/store/createStore';
 import reducer from './reducers';
 
 export default createStore(reducer);
