+++ script/library/core/module/GameObject.js
