+++ script/library/core/module/Sprite.js
