+++ script/library/core/module/SpriteFrame.js
