+++ script/library/core/module/SpriteSheet.js
