+++ script/library/core/utility/createEntity.js
